ml-service/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Data validation
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1

# ML dependencies - YOLO
ultralytics==8.3.51
torch>=2.0.0
torchvision>=0.15.0

# Image processing
opencv-python==4.10.0.84
Pillow==10.4.0
numpy>=1.24.0

# Utilities
psutil>=5.9.0  # For memory monitoring
